+++ ios/Classes/HyperSdkFlutterPlugin.m
@@ -1,11 +1,11 @@
 #import "HyperSdkFlutterPlugin.h"
-#if __has_include(<hypersdk/hypersdk-Swift.h>)
-#import <hypersdk/hypersdk-Swift.h>
+#if __has_include(<hypersdkflutter/hypersdkflutter-Swift.h>)
+#import <hypersdkflutter/hypersdkflutter-Swift.h>
 #else
 // Support project import fallback if the generated compatibility header
 // is not copied when this plugin is created as a library.
 // https://forums.swift.org/t/swift-static-libraries-dont-copy-generated-objective-c-header/19816
-#import "hypersdk-Swift.h"
+#import "hypersdkflutter-Swift.h"
 #endif
 
 @implementation HyperSdkFlutterPlugin